feat: implement show reward info for guest
