+++ lib/core/presentation/pages/event/event_detail_page/guest_event_detail_page/sub_pages/guest_event_reward_uses_page/guest_event_reward_uses_page.dart
@@ -1,6 +1,8 @@
 import 'package:app/core/application/auth/auth_bloc.dart';
 import 'package:app/core/application/event/get_event_detail_bloc/get_event_detail_bloc.dart';
 import 'package:app/core/application/event/get_event_reward_uses_bloc/get_event_reward_uses_bloc.dart';
+import 'package:app/core/domain/event/entities/event_reward_use.dart';
+import 'package:app/core/domain/event/entities/reward.dart';
 import 'package:app/core/domain/user/user_repository.dart';
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/sub_pages/guest_event_reward_uses_page/widgets/guest_event_reward_uses_listing.dart';
 import 'package:app/core/presentation/widgets/common/appbar/lemon_appbar_widget.dart';
@@ -52,6 +54,21 @@ class _GuestEventRewardUsesPageView extends StatelessWidget {
   Widget build(BuildContext context) {
     final t = Translations.of(context);
     final colorScheme = Theme.of(context).colorScheme;
+    String name = context.watch<AuthBloc>().state.maybeWhen(
+          authenticated: (authSession) => authSession.displayName ?? '',
+          orElse: () => "",
+        );
+
+    List<EventRewardUse>? eventRewardUses =
+        context.watch<GetEventRewardUsesBloc>().state.eventRewardUses ?? [];
+    List<Reward> eventRewards =
+        context.watch<GetEventDetailBloc>().state.maybeWhen(
+              fetched: (event) => event.rewards ?? [],
+              orElse: () => [],
+            );
+    int totalEventRewardUses = eventRewardUses.length;
+    int? totalLimitPer =
+        eventRewards.map((obj) => obj.limitPer).reduce((a, b) => a! + b!);
 
     return Scaffold(
       backgroundColor: colorScheme.background,
@@ -62,38 +79,33 @@ class _GuestEventRewardUsesPageView extends StatelessWidget {
         physics: const BouncingScrollPhysics(),
         slivers: [
           SliverAppBar(
-            leading: const SizedBox.shrink(),
-            collapsedHeight: kToolbarHeight,
-            pinned: true,
-            flexibleSpace: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                FutureBuilder(
-                  future: getIt<UserRepository>().getUser(userId: userId),
-                  builder: (context, snapshot) {
-                    final user =
-                        snapshot.data?.fold((l) => null, (user) => user);
-                    return Padding(
-                      padding: EdgeInsets.only(
-                        left: Spacing.smMedium,
-                        right: Spacing.smMedium,
-                        top: Spacing.xSmall,
+              leading: const SizedBox.shrink(),
+              collapsedHeight: kToolbarHeight + Spacing.superExtraSmall,
+              pinned: true,
+              flexibleSpace: Padding(
+                padding: EdgeInsets.symmetric(horizontal: Spacing.medium),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      t.event.rewards,
+                      style: Typo.superLarge.copyWith(
+                        color: colorScheme.onPrimary,
+                        fontWeight: FontWeight.w800,
+                        fontFamily: FontFamily.nohemiVariable,
                       ),
-                      child: Text(
-                        user?.name ?? '',
-                        style: Typo.superLarge.copyWith(
-                          color: colorScheme.onPrimary,
-                          fontWeight: FontWeight.w800,
-                          fontFamily: FontFamily.nohemiVariable,
-                        ),
+                    ),
+                    Text(
+                      '$totalEventRewardUses/$totalLimitPer ${t.common.claimed}',
+                      style: Typo.medium.copyWith(
+                        color: colorScheme.onSecondary,
+                        fontWeight: FontWeight.w800,
+                        fontFamily: FontFamily.nohemiVariable,
                       ),
-                    );
-                  },
+                    ),
+                  ],
                 ),
-              ],
-            ),
-          ),
+              )),
           BlocBuilder<GetEventRewardUsesBloc, GetEventRewardUsesState>(
             builder: (context, state) {
               if (state.initialLoading == true) {
@@ -104,7 +116,7 @@ class _GuestEventRewardUsesPageView extends StatelessWidget {
                   ),
                 );
               }
-              return GuestRewardUsesListing(userId: userId);
+              return const GuestEventRewardUsesListing();
             },
           ),
         ],
