+++ lib/core/presentation/pages/event/event_detail_page/host_event_detail_page/sub_pages/claim_rewards_page/widgets/claim_rewards_listing.dart
@@ -3,15 +3,62 @@ import 'package:app/core/application/event/get_event_reward_uses_bloc/get_event_
 import 'package:app/core/domain/event/entities/event.dart';
 import 'package:app/core/domain/event/entities/event_reward_use.dart';
 import 'package:app/core/domain/event/entities/reward.dart';
+import 'package:app/core/domain/event/repository/event_reward_repository.dart';
 import 'package:app/core/presentation/widgets/event/horizontal_rewards_list.dart';
+import 'package:app/core/presentation/widgets/future_loading_dialog.dart';
+import 'package:app/graphql/backend/schema.graphql.dart';
 import 'package:app/i18n/i18n.g.dart';
+import 'package:app/injection/register_module.dart';
 import 'package:app/theme/spacing.dart';
 import 'package:app/theme/typo.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_vibrate/flutter_vibrate.dart';
 
-class GuestEventRewardUsesListing extends StatelessWidget {
-  const GuestEventRewardUsesListing({super.key});
+class ClaimRewardsListing extends StatelessWidget {
+  const ClaimRewardsListing({super.key, required this.userId});
+  final String userId;
+
+  Future<void> onToggleClaim(
+    BuildContext context,
+    Reward reward,
+    int index,
+  ) async {
+    String eventId = context.watch<GetEventDetailBloc>().state.maybeWhen(
+          fetched: (event) => event.id ?? '',
+          orElse: () => '',
+        );
+    Vibrate.feedback(FeedbackType.light);
+    List<EventRewardUse>? eventRewardUses =
+        context.read<GetEventRewardUsesBloc>().state.eventRewardUses;
+    bool? exist = eventRewardUses?.any(
+      (item) => item.rewardNumber == index && item.rewardId == reward.id,
+    );
+    showFutureLoadingDialog(
+      context: context,
+      future: () async {
+        final result =
+            await getIt<EventRewardRepository>().updateEventRewardUse(
+          input: Input$UpdateEventRewardUseInput(
+            event: eventId,
+            user: userId,
+            reward_id: reward.id ?? '',
+            reward_number: index.toDouble(),
+            active: exist == false,
+          ),
+        );
+        if (result.isRight()) {
+          context.read<GetEventRewardUsesBloc>().add(
+                GetEventRewardUsesEvent.getEventRewardUses(
+                  eventId: eventId,
+                  userId: userId,
+                  showLoading: false,
+                ),
+              );
+        }
+      },
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -42,14 +89,12 @@ class GuestEventRewardUsesListing extends StatelessWidget {
                             )
                             .length ??
                         0;
-                    final leftCount = reward.limitPer! - totalCount;
                     return Padding(
                       padding: EdgeInsets.symmetric(
                         horizontal: Spacing.smMedium,
                         vertical: Spacing.small,
                       ),
                       child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
                           Text(
                             reward.title ?? '',
@@ -59,7 +104,7 @@ class GuestEventRewardUsesListing extends StatelessWidget {
                             width: Spacing.smMedium,
                           ),
                           Text(
-                            t.common.countLeft(count: leftCount),
+                            '$totalCount/${reward.limitPer} ${t.event.claimed}',
                             style: Typo.medium.copyWith(
                               color: colorScheme.onSecondary,
                             ),
@@ -70,8 +115,7 @@ class GuestEventRewardUsesListing extends StatelessWidget {
                   },
                 ),
                 HorizontalRewardsList(
-                  reward: reward,
-                ),
+                    reward: reward, onToggleItem: onToggleClaim),
               ],
             ),
           );
